UpperBitsNotSet assert in Chapter7 decoder, new Datatypes for smaller data and 32bit for enum implemented, Q_DIR is not in the header anymore

+++ Source/Adhesion_Factor/BothWays/BothWays_Header.h
@@ -6,7 +6,7 @@
 
 struct BothWays_Header
 {
-    uint64_t  NID_PACKET;         // # 8
+    uint8_t  NID_PACKET;         // # 8
 };
 
 typedef struct BothWays_Header BothWays_Header;
